Abstract progress bar code into its own package
