+++ common/go.mod
@@ -1,0 +1,21 @@
+module eth-wallet-gen/common
+
+go 1.17
+
+require (
+	github.com/cheggaaa/pb/v3 v3.0.8
+	github.com/schollz/progressbar/v3 v3.8.3
+)
+
+require (
+	github.com/VividCortex/ewma v1.1.1 // indirect
+	github.com/fatih/color v1.10.0 // indirect
+	github.com/mattn/go-colorable v0.1.8 // indirect
+	github.com/mattn/go-isatty v0.0.14 // indirect
+	github.com/mattn/go-runewidth v0.0.13 // indirect
+	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
+	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5 // indirect
+	golang.org/x/sys v0.0.0-20210910150752-751e447fb3d0 // indirect
+	golang.org/x/term v0.0.0-20210615171337-6886f2dfbf5b // indirect
+)
