+++ main.go
@@ -11,12 +11,12 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/cheggaaa/pb/v3"
 	hdwallet "github.com/miguelmota/go-ethereum-hdwallet"
-	"github.com/schollz/progressbar/v3"
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
+
+	"eth-wallet-gen/common"
 )
 
 var (
@@ -33,12 +33,6 @@ type Wallet struct {
 	HDPath     string
 	CreatedAt  time.Time
 	gorm.Model
-}
-
-//ProgressBar progressbar logging with 2 mode
-type ProgressBar struct {
-	StandardMode   *pb.ProgressBar
-	CompatibleMode *progressbar.ProgressBar
 }
 
 func generateNewWallet(bits int) *Wallet {
@@ -61,53 +55,6 @@ func createWallet(mnemonic string) *Wallet {
 		HDPath:     account.URL.Path,
 		CreatedAt:  time.Now(),
 	}
-}
-
-//NewProgressBar returns a new progress bar based on whether or not it's compatible.
-// A progress bar that is not compatible will be set to StandardMode while all
-// compatible progress bars will have their mode set to CompatibleMode.
-func NewProgressBar(number int, isCompatible bool) (bar *ProgressBar) {
-	if isCompatible {
-		return &ProgressBar{
-			CompatibleMode: progressbar.NewOptions(number,
-				progressbar.OptionSetItsString("w"),
-				progressbar.OptionSetPredictTime(true),
-				progressbar.OptionShowIts(),
-				progressbar.OptionShowCount(),
-				progressbar.OptionFullWidth(),
-			),
-		}
-	}
-	bar = &ProgressBar{
-		StandardMode: pb.StartNew(number),
-	}
-	bar.StandardMode.SetTemplate(pb.Default)
-	bar.StandardMode.SetTemplateString(`{{counters . }} | {{bar . "" "█" "█" "" "" | rndcolor}} | {{percent . }} | {{speed . }} | {{string . "resolved"}}`)
-	return
-}
-
-//Increment increment progress
-func (bar *ProgressBar) Increment() error {
-	if bar.CompatibleMode != nil {
-		return bar.CompatibleMode.Add(1)
-	}
-	return bar.StandardMode.Increment().Err()
-}
-
-//SetResolved set resolved wallet number
-func (bar *ProgressBar) SetResolved(resolved int) error {
-	if bar.StandardMode != nil {
-		return bar.StandardMode.Set("resolved", fmt.Sprintf("resolved: %v", resolved)).Err()
-	}
-	return nil
-}
-
-//Finish close progress bar
-func (bar *ProgressBar) Finish() error {
-	if bar.CompatibleMode != nil {
-		return bar.CompatibleMode.Close()
-	}
-	return bar.StandardMode.Finish().Err()
 }
 
 func init() {
@@ -201,7 +148,13 @@ func main() {
 		fmt.Printf("\nCopyright (C) 2021 Planxnx <planxthanee@gmail.com>\n")
 	}()
 
-	bar := NewProgressBar(*number, *isCompatible)
+	var bar *common.ProgressBar
+	if *isCompatible {
+		bar = common.NewCompatibleProgressBar(*number)
+	} else {
+		bar = common.NewStandardProgressBar(*number)
+	}
+
 	defer func() {
 		bar.Finish()
 		if *isDryrun {
